+++ tera/src/snapshot_tests/rendering.rs
@@ -5,10 +5,7 @@ use crate::delimiters::Delimiters;
 use crate::snapshot_tests::utils::{create_multi_templates_tera, normalize_line_endings};
 use crate::tera::Tera;
 
-#[cfg(not(feature = "preserve_order"))]
 use crate::args::Kwargs;
-
-#[cfg(not(feature = "preserve_order"))]
 use crate::vm::state::State;
 use crate::{Context, Value};
 
@@ -240,6 +237,17 @@ fn rendering_include_ok() {
     });
 }
 
+#[test]
+fn rendering_escaping_ok() {
+    insta::glob!("rendering_inputs/success/escaping/*.txt", |path| {
+        let contents = std::fs::read_to_string(path).unwrap();
+        let (tera, tpl_name) = create_multi_templates_tera(&contents);
+        let out = tera.render(&tpl_name, &get_context()).unwrap();
+        let normalized_out = normalize_line_endings(&out);
+        insta::assert_snapshot!(&normalized_out);
+    });
+}
+
 #[cfg(feature = "unicode")]
 #[test]
 fn can_iterate_on_graphemes() {
@@ -344,4 +352,27 @@ fn render_str_errors() {
 
     let out = tera.render_str(r#"{% include "missing.html" %}"#, &ctx, false);
     assert!(out.is_err());
+}
+
+#[test]
+fn render_capture_body_safety() {
+    let mut tera = Tera::default();
+    tera.autoescape_on(vec![".html"]);
+    tera.register_filter("is_safe", |s: Value, _: Kwargs, _: &State| s.is_safe());
+    tera.add_raw_templates(vec![
+        ("components.html", "{% component hey() %}a & b{% endcomponent %}{% component ho() %}{{body | is_safe }}{% endcomponent %}"),
+        ("a.txt", "{% set x %}a & b{% endset %}{{ x | is_safe }}-{{ <hey/> | is_safe }}-{% <ho> %}a & b{% </ho> %}"),
+        ("a.html", "{% set x %}a & b{% endset %}{{ x | is_safe }}-{{ <hey/> | is_safe }}-{% <ho> %}a & b{% </ho> %}"),
+    ]).unwrap();
+    let ctx = Context::default();
+
+    // It should only mark things as safe when autoescaping is enabled: capture, component output and component body
+    assert_eq!(
+        tera.render("a.txt", &ctx).unwrap().as_str(),
+        "false-false-false"
+    );
+    assert_eq!(
+        tera.render("a.html", &ctx).unwrap().as_str(),
+        "true-true-true"
+    );
 }