+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 2.3.0 (unreleased)
 
 - Add `State::{escape,autoescaping_enabled,escape_if_needed}`
+- Add `StringInput` to use for input/parameters that is safety aware and use it for built-in filters: this
+means you should need less `| safe` and that string safety is passed around correctly. Safety is dropped by
+filters where the changes depend on the content itself (`truncate` can cut anywhere) and preserved when they specify exactly what changes (`trim`, `replace`).
+- Add `escape` filter that is safety aware (it will not escape an already safe string) and uses the current instance
+escape function
+- Escaping included templates is now done base on _where_ it's included rather than the included template itself:
+for example a `.txt` included in a `.html` will now be autoescaped if Tera autoescapes on `.html`.
+- Fix double escape of filter/set blocks
 
 ## 2.2.0 (2026-08-20)